requirements.txt
requests~=2.32.3
beautifulsoup4~=4.13.3
matplotlib~=3.10.1
umap-learn~=0.5.7
streamlit
numpy~=1.24.4
sentence-transformers~=3.4.1
scikit-learn~=1.6.1
torch~=2.2.2
crossrefapi~=1.6.0
PyPDF2~=3.0.1